Post to lookups with static ip and newly created MAC
To resolve an issue where the template cannot be rendered because the information
is not in the lookups table.  This change will post to /lookups as the entries
are created, setting the static ip, mac, nodeid relationship in the database, so
the entry can be found when the template is requested to signal the completion of
the kickstart.